Update squat functionality to address bugs (#400)
Remove SeparatorRemovedMutator from the list of BaseMutators as it was causing problems and usually doesn't catch squats. Squats that would be caught by it are also caught by the prefix or suffix mutator.
Overrode the SeparatorChangedMutator for PyPi as PyPi considers - and _ to be the same character. So just removed _ from the list.
Fixed issues with the DuplicatorMutator that was giving the original name back as a mutated one.